requirements.txt
requests==2.31.0
pandas==2.0.3
pyodbc==5.1.1
python-dateutil==2.8.2
